requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-import-export==3.2.0
django-stubs==4.2.3
django-stubs-ext==4.2.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
mypy==1.4.1
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.11
typing_extensions==4.6.3
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
